``` Add rental association to Book and Person classes.
This commit adds a rental association to the Book and Person classes.
The Book class now has a rentals attribute that is an array of Rental objects.
The Person class now has a rentals attribute that is an array of Rental objects.
The Rental class has a book attribute that is a Book object and a person attribute that is a Person object.